copier: remove libimage dependency via util package
This allows podman to import the copier package on the remote client
without needing the full libimage package.

Based on Miloslav's work: https://github.com/containers/podman/pull/19718
