+++ copier/hardlink_not_uint64.go
@@ -1,6 +1,7 @@
+//go:build darwin || (linux && mips) || (linux && mipsle) || (linux && mips64) || (linux && mips64le)
 // +build darwin linux,mips linux,mipsle linux,mips64 linux,mips64le
 
-package util
+package copier
 
 import (
 	"syscall"
