+++ util/util_windows.go
@@ -1,3 +1,4 @@
+//go:build !linux && !darwin
 // +build !linux,!darwin
 
 package util
@@ -5,15 +6,6 @@ package util
 import (
 	"os"
 )
-
-type HardlinkChecker struct {
-}
-
-func (h *HardlinkChecker) Check(fi os.FileInfo) string {
-	return ""
-}
-func (h *HardlinkChecker) Add(fi os.FileInfo, name string) {
-}
 
 func UID(st os.FileInfo) int {
 	return 0
